pyproject.toml
[project]
name = "databot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.45.2",
    "chatlas>=0.3.0",
    "matplotlib>=3.10.0",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "polars>=1.22.0",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "shiny>=1.2.1",
]
